# Speedup for transform