This is a speedup for transformations like rotation, lensfun distortion correction...
I measured using a D800 file in full processing

1) only 1 degree rotation applied: before 400 ms, after 275 ms
2) 1 degree rotation + lensfun distortion correction: before 700 ms, after 570 ms

No differences in output